requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
distro==1.9.0
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.10
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.21
langchain-openai==0.2.11
langchain-upstage==0.4.0
langsmith==0.1.147
openai==1.56.2
orjson==3.10.12
packaging==24.2
pydantic==2.10.3
pydantic_core==2.27.1
pypdf==4.3.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.19.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
distro==1.9.0
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.10
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.21
langchain-openai==0.2.11
langchain-upstage==0.4.0
langsmith==0.1.147
openai==1.56.2
orjson==3.10.12
packaging==24.2
pydantic==2.10.3
pydantic_core==2.27.1
pypdf==4.3.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.19.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3